ref: 責務に応じた分割
