+++ src/config.py
@@ -10,7 +10,7 @@ def get_env_variable(name: str) -> str:
     return value
 
 
-env_path = os.path.join(os.path.dirname(__file__), '.env')
+env_path = os.path.join(os.path.dirname(__file__), '..', '.env')
 load_dotenv(dotenv_path=env_path)
 
 # --- Misskey API関連 ---
@@ -32,4 +32,4 @@ MAX_NICKNAME_LENGTH = 15
 
 # --- ファイルパス ---
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-NICKNAME_FILE = os.path.join(BASE_DIR, "nickname.json")
+NICKNAME_FILE = os.path.join(BASE_DIR, '..', 'nickname.json')
