+++ src/utils.py
@@ -40,3 +40,10 @@ def validate_nickname(name: str) -> bool:
     if name.isspace():
         return False
     return True
+
+
+def create_mention_string(user: dict) -> str:
+    """@username または @username@host 形式のメンション文字列を生成する"""
+    username = user.get("username")
+    host = user.get("host")
+    return f"@{username}@{host}" if host else f"@{username}"
